% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{check_answer}
\alias{check_answer}
\title{Check if a single question is answered}
\usage{
check_answer(q, input)
}
\arguments{
\item{q}{Question identifier}

\item{input}{Shiny input object}
}
\value{
Logical indicating if the question is answered
}
\description{
Check if a single question is answered
}
\keyword{internal}
