% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_skip_show}
\alias{check_skip_show}
\title{Check skip and show conditions}
\usage{
check_skip_show(
  question_ids,
  question_values,
  page_ids,
  skip_if,
  skip_if_custom,
  show_if,
  show_if_custom
)
}
\arguments{
\item{skip_if}{Data frame of skip conditions}

\item{skip_if_custom}{Custom skip function}

\item{show_if}{Data frame of show conditions}

\item{show_if_custom}{Custom show function}

\item{config}{The survey configuration list}
}
\value{
TRUE if all checks pass, otherwise stops with an error message
}
\description{
This function validates the skip_if and show_if conditions provided in the configuration.
}
\keyword{internal}
