% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{check_surveydown_versions}
\alias{check_surveydown_versions}
\title{Check Surveydown Versions}
\usage{
check_surveydown_versions()
}
\value{
No return value, called for side effects (prints version information).
}
\description{
This function checks the versions of both the surveydown R package and its
associated Quarto extension, reporting any mismatches.
}
\examples{
check_surveydown_versions()
}
