% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_table}
\alias{create_table}
\title{Create a new table in the database}
\usage{
create_table(db, table_name, df)
}
\arguments{
\item{db}{Database connection object}

\item{table_name}{String name of the table to create}

\item{df}{Data frame used to determine table structure}
}
\value{
None (called for side effects)
}
\description{
Create a new table in the database
}
\keyword{internal}
