% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{database_uploading}
\alias{database_uploading}
\title{Upload data to the database}
\usage{
database_uploading(df, db, table_name)
}
\arguments{
\item{df}{Data frame of survey data to upload}

\item{db}{Database connection object}

\item{table_name}{String name of the table to upload to}
}
\value{
None (called for side effects)
}
\description{
This function handles the process of uploading survey data to the database,
including creating the table if it doesn't exist and updating existing rows.
}
\keyword{internal}
