% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{download_extension}
\alias{download_extension}
\title{Download and Extract Survey Extension}
\usage{
download_extension(temp_dir)
}
\arguments{
\item{temp_dir}{A temporary directory to store downloaded files.}
}
\value{
A character string with the path to the unzipped directory.
}
\description{
This helper function downloads the surveydown extension from GitHub and extracts it.
}
\keyword{internal}
