% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_extension_version}
\alias{get_extension_version}
\title{Get Surveydown Extension Version}
\usage{
get_extension_version(path = getwd())
}
\arguments{
\item{path}{A character string specifying the directory to search for the
extension. Defaults to the current working directory.}
}
\value{
A character string representing the extension version, or NULL if
the extension is not found.
}
\description{
This function reads the version of the surveydown Quarto extension from its
_extension.yml file.
}
\keyword{internal}
