% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_latest_version_from_url}
\alias{get_latest_version_from_url}
\title{Get Latest Version from URL}
\usage{
get_latest_version_from_url(url, pattern)
}
\arguments{
\item{url}{The URL of the file containing the version information}

\item{pattern}{The pattern to search for in the file (e.g., "Version: " or "version: ")}
}
\value{
A package_version object representing the latest version, or NULL if unable to fetch
}
\description{
This function fetches the latest version from a file at a given URL.
}
\keyword{internal}
