% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_page_nodes}
\alias{get_page_nodes}
\title{Get page nodes from HTML}
\usage{
get_page_nodes()
}
\value{
A list of page nodes from the HTML document
}
\description{
Get page nodes from HTML
}
\keyword{internal}
