% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_page_structure}
\alias{get_page_structure}
\title{Get page structure from HTML}
\usage{
get_page_structure()
}
\value{
A list where each element represents a page and contains the question IDs on that page
}
\description{
Get page structure from HTML
}
\keyword{internal}
