% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_question_nodes}
\alias{get_question_nodes}
\title{Get question nodes from HTML}
\usage{
get_question_nodes()
}
\value{
A list of question nodes from the HTML document
}
\description{
Get question nodes from HTML
}
\keyword{internal}
