% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_question_structure}
\alias{get_question_structure}
\title{Get question structure from HTML}
\usage{
get_question_structure()
}
\value{
A list where each element represents a question and contains its options
}
\description{
Get question structure from HTML
}
\keyword{internal}
