% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_utc_timestamp}
\alias{get_utc_timestamp}
\title{Get Current UTC Timestamp}
\usage{
get_utc_timestamp()
}
\value{
A character string representing the current UTC timestamp.
}
\description{
This function returns the current timestamp in UTC format.
}
\keyword{internal}
