% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_basic_show_if_logic}
\alias{handle_basic_show_if_logic}
\title{Handle basic show-if logic}
\usage{
handle_basic_show_if_logic(input, show_if)
}
\arguments{
\item{input}{Shiny input object}

\item{show_if}{Data frame of show-if conditions}
}
\description{
Handle basic show-if logic
}
\keyword{internal}
