% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_basic_skip_logic}
\alias{handle_basic_skip_logic}
\title{Handle basic skip logic}
\usage{
handle_basic_skip_logic(input, skip_if, current_page, next_page)
}
\arguments{
\item{input}{Shiny input object}

\item{skip_if}{Data frame of skip-if conditions}

\item{current_page}{Current page identifier}

\item{next_page}{Next page identifier}
}
\value{
Updated next page identifier
}
\description{
Handle basic skip logic
}
\keyword{internal}
