% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_custom_show_if_logic}
\alias{handle_custom_show_if_logic}
\title{Handle custom show-if logic}
\usage{
handle_custom_show_if_logic(input, show_if_custom)
}
\arguments{
\item{input}{Shiny input object}

\item{show_if_custom}{List of custom show-if conditions}
}
\description{
Handle custom show-if logic
}
\keyword{internal}
