% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_custom_skip_logic}
\alias{handle_custom_skip_logic}
\title{Handle custom skip logic}
\usage{
handle_custom_skip_logic(input, skip_if_custom, current_page, next_page)
}
\arguments{
\item{input}{Shiny input object}

\item{skip_if_custom}{List of custom skip-if conditions}

\item{current_page}{Current page identifier}

\item{next_page}{Next page identifier}
}
\value{
Updated next page identifier
}
\description{
Handle custom skip logic
}
\keyword{internal}
