% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{hide_all_pages}
\alias{hide_all_pages}
\title{Hide All Survey Pages}
\usage{
hide_all_pages()
}
\value{
None (invisible NULL)
}
\description{
This function runs a JavaScript code snippet that hides all elements
with the class 'sd-page'.
}
\keyword{internal}
