% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{hide_sd_pages}
\alias{hide_sd_pages}
\title{Hide All Elements with the .sd-page Class}
\usage{
hide_sd_pages()
}
\description{
This function hides all HTML elements on the page that have the \code{.sd-page} class.
It uses JavaScript to set the \code{display} style of these elements to \code{'none'}, effectively making them invisible on the webpage.
}
\examples{
\dontrun{
hide_sd_pages()
}
}
