% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{initialize_timestamps}
\alias{initialize_timestamps}
\title{Initialize Timestamps for Pages and Questions}
\usage{
initialize_timestamps(page_ids, question_ids)
}
\arguments{
\item{page_ids}{A vector of page IDs.}

\item{question_ids}{A vector of question IDs.}
}
\value{
A list of initialized timestamps for pages and questions.
}
\description{
This function initializes timestamps for pages and questions in a survey.
}
\keyword{internal}
