% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{is_question_visible}
\alias{is_question_visible}
\title{Check if a question is visible}
\usage{
is_question_visible(q, show_if, input)
}
\arguments{
\item{q}{Question identifier}

\item{show_if}{Data frame of show-if conditions}

\item{input}{Shiny input object}
}
\value{
Logical indicating if the question is visible
}
\description{
Check if a question is visible
}
\keyword{internal}
