% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list_name_md_to_html}
\alias{list_name_md_to_html}
\title{Convert List Names from Markdown to HTML}
\usage{
list_name_md_to_html(list)
}
\arguments{
\item{list}{A list whose names are in markdown format.}
}
\value{
A list with names converted to HTML format.
}
\description{
This function converts the names of a list from markdown to HTML format.
It only works for mc_buttons and mc_multiple_buttons.
}
\keyword{internal}
