% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{make_next_button_id}
\alias{make_next_button_id}
\title{Generate Next Button ID}
\usage{
make_next_button_id(next_page)
}
\arguments{
\item{next_page}{Character string. The ID of the next page.}
}
\value{
A character string representing the button ID.
}
\description{
This internal function generates a unique ID for the 'Next' button based on the next page ID.
}
\keyword{internal}
