% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_ts_name}
\alias{make_ts_name}
\title{Make Timestamp Name}
\usage{
make_ts_name(type, id)
}
\arguments{
\item{type}{A character string, either "page" or "question".}

\item{id}{The ID of the page or question.}
}
\value{
A character string representing the timestamp name.
}
\description{
This function creates a standardized name for timestamps based on the type and ID.
}
\keyword{internal}
