% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{markdown_to_html}
\alias{markdown_to_html}
\title{Convert Markdown to HTML}
\usage{
markdown_to_html(text)
}
\arguments{
\item{text}{A character string containing markdown text.}
}
\value{
An HTML string converted from the input markdown text.
}
\description{
This function converts markdown text to HTML.
}
\keyword{internal}
