% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{r_to_sql_type}
\alias{r_to_sql_type}
\title{Convert R data type to SQL data type}
\usage{
r_to_sql_type(r_type)
}
\arguments{
\item{r_type}{String representing R data type}
}
\value{
String representing corresponding SQL data type
}
\description{
Convert R data type to SQL data type
}
\keyword{internal}
