% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sd_admin}
\alias{sd_admin}
\title{Launch Admin Interface}
\usage{
sd_admin(table)
}
\arguments{
\item{table}{Character string. The name of the table to interact with.}
}
\value{
No return value, called for side effects (opens Shiny gadget)
}
\description{
Opens a Shiny gadget that provides administrative functionality for surveydown surveys,
including viewing survey responses and managing survey data.
}
