% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_admin_ui}
\alias{sd_admin_ui}
\title{Custom Admin UI Function}
\usage{
sd_admin_ui()
}
\value{
A shiny tagList containing the admin UI elements.
}
\description{
This function creates a custom admin UI for the surveydown package.
}
