% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sd_config}
\alias{sd_config}
\title{Configuration Function for surveydown Surveys}
\usage{
sd_config(
  use_html = FALSE,
  skip_if = NULL,
  skip_if_custom = NULL,
  show_if = NULL,
  show_if_custom = NULL,
  required_questions = NULL,
  all_questions_required = FALSE,
  start_page = NULL,
  admin_page = FALSE
)
}
\arguments{
\item{use_html}{Logical. By default, sd_config() will render the "survey.qmd" file when loaded, which can be slow. User can render it first into a html file and set \code{use_html = TRUE} to use the pre-rendered file, which is faster when the app loads. Defaults to \code{FALSE}.}

\item{skip_if}{A list of conditions under which certain pages should be skipped. Defaults to NULL.}

\item{skip_if_custom}{A custom function to handle conditions under which certain pages should be skipped. Defaults to NULL.}

\item{show_if}{A list of conditions under which certain pages should be shown. Defaults to NULL.}

\item{show_if_custom}{A custom function to handle conditions under which certain pages should be shown. Defaults to NULL.}

\item{required_questions}{Vector of character strings. The IDs of questions that must be answered. Defaults to NULL.}

\item{all_questions_required}{Logical. If TRUE, all questions in the survey will be required. Defaults to FALSE.}

\item{start_page}{Character string. The ID of the page to start on. Defaults to NULL.}

\item{admin_page}{Logical. Whether to include an admin page for viewing and downloading survey data. Defaults to FALSE.}
}
\value{
A list containing the configuration settings for the survey and rendered HTML content.
}
\description{
This function sets up the configuration for a surveydown survey, including
page and question structures, conditional display settings, and navigation options.
It also renders the Quarto document and extracts necessary information.
}
