% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_display_value}
\alias{sd_display_value}
\title{Display the value of a survey question}
\usage{
sd_display_value(id, display_type = "inline", wrapper = NULL, ...)
}
\arguments{
\item{id}{The ID of the question to display}

\item{display_type}{The type of display. Can be "inline" (default), "text", "verbatim", or "ui".}

\item{wrapper}{A function to wrap the output}

\item{...}{Additional arguments passed to the wrapper function}
}
\value{
A Shiny UI element displaying the question's value
}
\description{
Display the value of a survey question
}
\examples{
sd_display_value("name")
sd_display_value("age", display_type = "text")
\dontrun{
  sd_display_value("email", display_type = "inline", wrapper = function(x) tags$strong(x))
}

}
