% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_filter}
\alias{sd_filter}
\title{Set up Surveydown Filter for Quarto}
\usage{
sd_filter()
}
\value{
Invisibly returns the path to the Surveydown Lua filter.
}
\description{
This function sets up the Surveydown Lua filter for use in Quarto documents.
It finds the filter path and sets the appropriate Quarto option.
}
\examples{
\dontrun{
sd_filter()
}
}
