% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_page_gadget}
\alias{sd_page_gadget}
\title{Show a Shiny gadget for entering a page ID}
\usage{
sd_page_gadget()
}
\value{
The entered page ID (invisibly).
}
\description{
This function displays a Shiny gadget that allows the user to input
a page ID. Once submitted, it calls sd_add_page() with the specified ID.
}
