% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_question_gadget}
\alias{sd_question_gadget}
\title{Show a Shiny gadget for selecting a question type}
\usage{
sd_question_gadget(chunk = FALSE)
}
\arguments{
\item{chunk}{Logical. If \code{TRUE}, the code will be generated with the R code
chunk wrapper. Defaults to \code{FALSE}.}
}
\value{
The selected question type (invisibly).
}
\description{
This function displays a Shiny gadget that allows the user to select
a question type from a dropdown menu. Once submitted, it calls
sd_add_question() with the specified type.
}
