% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_reactive_output}
\alias{sd_reactive_output}
\title{Create a placeholder for a reactive survey question}
\usage{
sd_reactive_output(id)
}
\arguments{
\item{id}{A unique identifier for the question.}
}
\value{
A Shiny UI element that serves as a placeholder for the reactive question.
}
\description{
This function creates a placeholder div for a reactive survey question.
}
