% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_render_url}
\alias{sd_render_url}
\title{Render URL Output in Shiny}
\usage{
sd_render_url(outputId, ...)
}
\arguments{
\item{outputId}{A character string specifying the output ID for the UI element.}

\item{...}{Additional arguments passed to \code{shiny::uiOutput}.}
}
\value{
A UI output element.
}
\description{
A wrapper function for \code{shiny::uiOutput}, used for consistency in URL-related render functions.
}
