% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sd_set_password}
\alias{sd_set_password}
\title{Set Supabase Password}
\usage{
sd_set_password(password)
}
\arguments{
\item{password}{Character string. The password to be set for Supabase connection.}
}
\value{
None. The function is called for its side effects.
}
\description{
This function sets the supabase password in the .Renviron file and adds .Renviron to .gitignore.
}
\details{
The function performs the following actions:
\enumerate{
\item Creates a .Renviron file in the root directory if it doesn't exist.
\item Adds or updates the SURVEYDOWN_PASSWORD entry in the .Renviron file.
\item Adds .Renviron to .gitignore if it's not already there.
}
}
\examples{
\dontrun{
  sd_set_password("your_SURVEYDOWN_PASSWORD")
}

}
