% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_setup}
\alias{sd_setup}
\title{Required Set Up Function}
\usage{
sd_setup()
}
\value{
This function does not return a value. It is called for its side effects of setting up the Shiny application.
}
\description{
This function is required for any surveydown survey. It sets up a Shiny application with Bootstrap 5 and initializes Shinyjs for JavaScript functionalities.
}
\details{
The function configures the Shiny application to use Bootstrap 5 for styling and enables
Shinyjs for JavaScript functionalities within the application.
}
\examples{
\dontrun{
  ui <- fluidPage(
    sd_setup(),
    # Your UI elements here
  )
  server <- function(input, output, session) {
    # Your server logic here
  }
  shinyApp(ui, server)
}

}
