% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_show_if}
\alias{sd_show_if}
\title{Define show conditions for survey questions}
\usage{
sd_show_if(...)
}
\arguments{
\item{...}{One or more formulas defining show conditions.
The left-hand side of each formula should be a condition based on input values,
and the right-hand side should be the ID of the question to show if the condition is met.}
}
\value{
A list of parsed conditions, where each element contains the condition and the target question ID.
Returns NULL if no conditions are provided.
}
\description{
This function is used to define conditions under which certain questions in the survey should be shown.
It takes one or more formulas where the left-hand side is the condition and the right-hand side is the target question ID.
If called with no arguments, it will return NULL and set no conditions.
}
\examples{
\dontrun{
sd_show_if(
  input$has_pets == "yes" ~ "pet_details",
  input$employment == "employed" ~ "job_questions"
)
}

}
\seealso{
\code{\link{sd_skip_if}}
}
