% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_show_pages}
\alias{sd_show_pages}
\title{Show Only the First Element with the .sd-page Class}
\usage{
sd_show_pages()
}
\description{
This function shows only the first HTML element on the page that has the \code{.sd-page} class.
It uses JavaScript to set the \code{display} style of the first such element to \code{'block'}, making it visible on the webpage.
}
\examples{
\dontrun{
sd_show_pages()
}
}
