% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sd_update}
\alias{sd_update}
\title{Update Surveydown Package}
\usage{
sd_update(force = FALSE)
}
\arguments{
\item{force}{Logical; if TRUE, forces an update regardless of current version.
Defaults to FALSE.}
}
\value{
No return value, called for side effects.
}
\description{
This function checks and updates surveydown.
It ensures that the package is up-to-date.
}
\examples{
\dontrun{
sd_update()
sd_update(force = TRUE)
}
}
