% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sd_update_surveydown}
\alias{sd_update_surveydown}
\title{Update Surveydown Package and Extension}
\usage{
sd_update_surveydown(force = FALSE)
}
\arguments{
\item{force}{Logical; if TRUE, forces an update regardless of current versions.
Defaults to FALSE.}
}
\value{
No return value, called for side effects.
}
\description{
This function checks and updates both the surveydown R package and its
associated Quarto extension. It ensures that both components are up-to-date
and their versions match.
}
\examples{
\dontrun{
sd_update_surveydown()
sd_update_surveydown(force = TRUE)
}
}
