% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_version}
\alias{sd_version}
\title{Check Surveydown Version}
\usage{
sd_version()
}
\value{
No return value, called for side effects (prints version information
and update status to the console).
}
\description{
This function checks if the local surveydown package is up-to-date with
the latest online version. It compares the local version with the latest
version available on GitHub and provides information about whether an update
is needed.
}
\examples{
sd_version()
}
