% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{show_first_page}
\alias{show_first_page}
\title{Show First Survey Page}
\usage{
show_first_page()
}
\value{
None (invisible NULL)
}
\description{
This function runs a JavaScript code snippet that displays the first
element with the class 'sd-page'.
}
\keyword{internal}
