% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{update_extension}
\alias{update_extension}
\title{Update Survey Extension}
\usage{
update_extension(path = getwd())
}
\arguments{
\item{path}{A character string specifying the directory in which to update
or create the extension. Defaults to the current working directory.}
}
\value{
A message indicating the successful update of the extension.
}
\description{
This function updates or creates the _extensions/jhelvy/surveydown folder
with the latest contents from the surveydown-ext repository.
}
\examples{
\dontrun{
update_extension()
update_extension(path = "path/to/project")
}
}
