% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vector_to_json_array}
\alias{vector_to_json_array}
\title{Convert Vector to JSON Array}
\usage{
vector_to_json_array(vec)
}
\arguments{
\item{vec}{A vector to be converted to a JSON array.}
}
\value{
A character string representing the JSON array.
}
\description{
This function converts a vector to a JSON array string. It's a custom alternative
to jsonlite::toJSON().
}
\keyword{internal}
