% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_add_admin_functionality}
\alias{sd_add_admin_functionality}
\title{Add Admin Functionality to Survey}
\usage{
sd_add_admin_functionality(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\description{
This function adds admin functionality to a surveydown survey, including
an admin button, login page, and basic admin page.
}
