% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{sd_admin}
\alias{sd_admin}
\title{Launch Admin App}
\usage{
sd_admin()
}
\value{
No return value, called for side effects (opens admin interface)
}
\description{
Opens an admin interface to view survey responses and summary statistics
}
