% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sd_check_versions}
\alias{sd_check_versions}
\title{Check Surveydown Versions}
\usage{
sd_check_versions()
}
\value{
No return value, called for side effects (prints version information).
}
\description{
This function checks the versions of both the surveydown R package and its
associated Quarto extension, reporting any mismatches and suggesting an update
if necessary.
}
\examples{
sd_check_versions()
}
