% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sd_check_versions}
\alias{sd_check_versions}
\title{Check Surveydown Versions}
\usage{
sd_check_versions()
}
\value{
No return value, called for side effects (prints version information).
}
\description{
This function checks if the local surveydown R package and Quarto extension
are up-to-date with the latest online version.
}
\examples{
sd_check_versions()
}
