% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_close}
\alias{sd_close}
\title{Create a 'Close' Button to Exit the Survey}
\usage{
sd_close(label = "Close")
}
\arguments{
\item{label}{Character string. The label of the 'Close' button. Defaults to "Close".}
}
\value{
A Shiny action button UI element with associated JavaScript for closing the page and Enter key functionality.
}
\description{
This function creates a 'Close' button that, when clicked, will close the current browser tab or window.
The button can be activated by clicking or by pressing the Enter key.
}
\details{
The function generates a Shiny action button that, when clicked or when the Enter key is pressed,
will attempt to close the current browser tab or window. Note that for security reasons,
some browsers may not allow JavaScript to close windows that were not opened by JavaScript.
In such cases, the button will prompt the user to close the tab manually.
}
\examples{
sd_close()
sd_close("Exit Survey")

}
