% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_completion_code}
\alias{sd_completion_code}
\title{Generate a Random Completion Code}
\usage{
sd_completion_code(digits = 6)
}
\arguments{
\item{digits}{An integer specifying the number of digits in the completion code.
Must be a positive integer. Default is 6.}
}
\value{
A character string representing the random completion code.
}
\description{
This function generates a random completion code with a specified number of digits.
The code is returned as a character string.
}
\examples{
sd_completion_code()  # generates a 6-digit code
sd_completion_code(digits = 8)  # generates an 8-digit code
sd_completion_code(digits = 4)  # generates a 4-digit code
sd_completion_code(digits = 10)  # generates a 10-digit code

}
