% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sd_config}
\alias{sd_config}
\title{Configuration Function for surveydown Surveys}
\usage{
sd_config(
  skip_if = NULL,
  skip_if_custom = NULL,
  show_if = NULL,
  show_if_custom = NULL,
  preview = FALSE,
  start_page = NULL,
  show_all_pages = FALSE
)
}
\arguments{
\item{skip_if}{A list of conditions under which certain pages should be skipped. Defaults to NULL.}

\item{skip_if_custom}{A custom function to handle conditions under which certain pages should be skipped. Defaults to NULL.}

\item{show_if}{A list of conditions under which certain pages should be shown. Defaults to NULL.}

\item{show_if_custom}{A custom function to handle conditions under which certain pages should be shown. Defaults to NULL.}

\item{preview}{Logical. Whether the survey is in preview mode. Defaults to FALSE.}

\item{start_page}{Character string. The ID of the page to start on. Defaults to NULL.}

\item{show_all_pages}{Logical. Whether to show all pages initially. Defaults to FALSE.}
}
\value{
A list containing the configuration settings for the survey, including page and question
structures, conditional display settings, and navigation options.
}
\description{
This function sets up the configuration for a surveydown survey, including
page and question structures, conditional display settings, and navigation options.
}
\details{
The function retrieves the survey metadata, checks the validity of the conditional
display settings, and ensures that the specified start page (if any) exists. It then stores
these settings in a configuration list.
}
\examples{
\dontrun{
  config <- sd_config(
    skip_if = list(),
    skip_if_custom = NULL,
    show_if = list(),
    show_if_custom = NULL,
    preview = FALSE,
    start_page = "page1",
    show_all_pages = FALSE
  )
}

}
