% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sd_config}
\alias{sd_config}
\title{Configuration Function for surveydown Surveys}
\usage{
sd_config(
  skip_if = NULL,
  skip_if_custom = NULL,
  show_if = NULL,
  show_if_custom = NULL,
  required_questions = NULL,
  start_page = NULL,
  show_all_pages = FALSE,
  admin_page = FALSE
)
}
\arguments{
\item{skip_if}{A list of conditions under which certain pages should be skipped. Defaults to NULL.}

\item{skip_if_custom}{A custom function to handle conditions under which certain pages should be skipped. Defaults to NULL.}

\item{show_if}{A list of conditions under which certain pages should be shown. Defaults to NULL.}

\item{show_if_custom}{A custom function to handle conditions under which certain pages should be shown. Defaults to NULL.}

\item{required_questions}{Vector of character strings. The IDs of questions that must
be answered before the respondent can continue in the survey or survey can be
submitted. Defaults to NULL.}

\item{start_page}{Character string. The ID of the page to start on. Defaults to NULL.}

\item{show_all_pages}{Logical. Whether to show all pages initially. Defaults to FALSE.}

\item{admin_page}{Logical. Whether to include an admin page for viewing and downloading survey data. Defaults to FALSE.}
}
\value{
A list containing the configuration settings for the survey, including page and question
structures, conditional display settings, navigation options, and admin page settings.
}
\description{
This function sets up the configuration for a surveydown survey, including
page and question structures, conditional display settings, and navigation options.
}
\details{
The function retrieves the survey metadata, checks the validity of the conditional
display settings, and ensures that the specified start page (if any) exists. It then stores
these settings in a configuration list. If \code{admin_page} is set to TRUE, an admin page will be
included in the survey. This page allows viewing and downloading of survey data upon entering
the correct survey password (set using \code{sd_set_password()}).
}
\examples{
\dontrun{
  config <- sd_config(
    skip_if        = list(),
    skip_if_custom = NULL,
    show_if        = list(),
    show_if_custom = NULL,
    start_page     = "page1",
    show_all_pages = FALSE,
    admin_page     = TRUE
  )
}

}
