% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sd_config}
\alias{sd_config}
\title{Configuration Function for surveydown Surveys}
\usage{
sd_config(
  survey_qmd = "survey.qmd",
  skip_if = NULL,
  skip_if_custom = NULL,
  show_if = NULL,
  show_if_custom = NULL,
  required_questions = NULL,
  all_questions_required = FALSE,
  start_page = NULL,
  show_all_pages = FALSE,
  admin_page = FALSE
)
}
\arguments{
\item{survey_qmd}{Character string. The path to the survey .qmd file.}

\item{skip_if}{A list of conditions under which certain pages should be skipped. Defaults to NULL.}

\item{skip_if_custom}{A custom function to handle conditions under which certain pages should be skipped. Defaults to NULL.}

\item{show_if}{A list of conditions under which certain pages should be shown. Defaults to NULL.}

\item{show_if_custom}{A custom function to handle conditions under which certain pages should be shown. Defaults to NULL.}

\item{required_questions}{Vector of character strings. The IDs of questions that must be answered. Defaults to NULL.}

\item{all_questions_required}{Logical. If TRUE, all questions in the survey will be required. Defaults to FALSE.}

\item{start_page}{Character string. The ID of the page to start on. Defaults to NULL.}

\item{show_all_pages}{Logical. Whether to show all pages initially. Defaults to FALSE.}

\item{admin_page}{Logical. Whether to include an admin page for viewing and downloading survey data. Defaults to FALSE.}
}
\value{
A list containing the configuration settings for the survey and rendered HTML content.
}
\description{
This function sets up the configuration for a surveydown survey, including
page and question structures, conditional display settings, and navigation options.
It also renders the Quarto document and extracts necessary information.
}
