% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sd_connect}
\alias{sd_connect}
\title{Database connection function}
\usage{
sd_connect(table = NULL, ignore = FALSE)
}
\arguments{
\item{table}{table name}

\item{ignore}{if TRUE, use local csv instead}
}
\value{
connection list or NULL
}
\description{
reads configuration file and establishes database connection
}
