% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_copy_value}
\alias{sd_copy_value}
\title{Create a copy of an input value}
\usage{
sd_copy_value(id, id_copy)
}
\arguments{
\item{id}{The ID of the input value to copy}

\item{id_copy}{The ID for the new copy (must be different from id)}
}
\value{
NULL invisibly. This function is called for its side effects.
}
\description{
This function creates a copy of an input value and makes it available as a new output.
The new output can then be displayed using sd_display_value().
}
\examples{
\dontrun{
sd_copy_value(id = "respondent_name", id_copy = "resp_name2")

# Then in UI:
# sd_display_value("resp_name2")
}

}
