% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_create_survey}
\alias{sd_create_survey}
\title{Create a new survey template}
\usage{
sd_create_survey(path = getwd())
}
\arguments{
\item{path}{A character string specifying the directory where the survey
template should be created. Defaults to the current working directory.}
}
\value{
Invisible \code{NULL}. The function is called for its side effects.
}
\description{
Create a new survey template
}
