% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sd_database}
\alias{sd_database}
\title{Connect to a SupaBase Database}
\usage{
sd_database(host, db_name, port, user, table_name, password)
}
\arguments{
\item{host}{Character string. The host address of the SupaBase database.}

\item{db_name}{Character string. The name of the SupaBase database.}

\item{port}{Integer. The port number for the SupaBase database connection.}

\item{user}{Character string. The username for the SupaBase database connection.}

\item{table_name}{Character string. The name of the table to interact with in the SupaBase database.}

\item{password}{Character string. The password for the SupaBase database connection.}
}
\value{
A list containing the database connection object (\code{db}) and the table name (\code{table_name}).
}
\description{
This function establishes a connection to a SupaBase database using the provided
connection details.
}
\details{
The function checks for the presence of all required parameters and attempts to
establish a connection to the SupaBase database. If successful, it returns a list containing
the database connection object and the table name. The user must have created the specified
table in SupaBase beforehand.
}
\note{
The user must create their own table inside SupaBase in order to make additions.
}
\examples{
\dontrun{
  db_connection <- sd_database(
    host = "your-host",
    db_name = "your-db-name",
    port = 5432,
    user = "your-username",
    table_name = "your-table-name",
    password = "your-password"
  )
}

}
