% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sd_db_config}
\alias{sd_db_config}
\title{Configure Database Settings}
\usage{
sd_db_config()
}
\value{
Invisibly returns a list of the configured settings if successful, or NULL if cancelled.
}
\description{
Opens a user interface to configure database connection settings for surveydown.
These settings are saved in a .env file and are used for storing survey responses.
The interface allows for testing the connection before saving settings.
}
\examples{
if (interactive()) {
  # Open the configuration interface
  sd_db_config()
}

}
