% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_deploy}
\alias{sd_deploy}
\title{Deploy a Surveydown App}
\usage{
sd_deploy(name = "survey")
}
\arguments{
\item{name}{A character string specifying the name of the app. Default is "survey".}
}
\value{
This function doesn't return a value; it deploys the app to Shiny Server.
}
\description{
This function is a wrapper for \code{rsconnect::deployApp()} specifically designed
for deploying Surveydown applications. It simplifies the deployment process
by allowing you to specify just the app name.
}
\examples{
\dontrun{
# Deploy with default name "survey"
sd_deploy()

# Deploy with a custom name
sd_deploy("my_custom_survey")
}

}
\seealso{
\code{\link[rsconnect]{deployApp}}
}
