% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_display_question}
\alias{sd_display_question}
\title{Create a placeholder for a reactive survey question}
\usage{
sd_display_question(id)
}
\arguments{
\item{id}{A unique identifier for the question.}
}
\value{
A Shiny UI element that serves as a placeholder for the reactive question.
}
\description{
This function creates a placeholder div for a reactive survey question in a Surveydown survey.
It's used in conjunction with sd_question_reactive to allow for dynamic question rendering.
}
\examples{
sd_display_question("name")

}
