% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_get_url_pars}
\alias{sd_get_url_pars}
\title{Get URL Parameters in a Shiny Application}
\usage{
sd_get_url_pars(...)
}
\arguments{
\item{...}{Optional. Names of specific URL parameters to retrieve.
If none are specified, all URL parameters are returned.}
}
\value{
A reactive expression that returns a list of URL parameters.
}
\description{
This function retrieves URL parameters from the current Shiny session.
It must be called from within a Shiny reactive context.
}
\examples{
# Examples here
}
