% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_is_answered}
\alias{sd_is_answered}
\title{Check if a question is answered}
\usage{
sd_is_answered(question_id)
}
\arguments{
\item{question_id}{The ID of the question to check.}
}
\value{
A logical value: TRUE if the question is answered, FALSE otherwise.
}
\description{
This function checks if a given question has been answered by the user.
For matrix questions, it checks if all sub-questions are answered.
}
\examples{
\dontrun{
if(sd_is_answered("car_preference")) {
  # Do something when all sub-questions of the matrix are answered
}
}

}
