% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_is_answered}
\alias{sd_is_answered}
\title{Check if a question is answered}
\usage{
sd_is_answered(question_id)
}
\arguments{
\item{question_id}{The ID of the question to check.}
}
\value{
A logical value: \code{TRUE} if the question is answered, \code{FALSE}
otherwise.
}
\description{
This function checks if a given question has been answered by the user.
For matrix questions, it checks if all sub-questions (rows) are answered.
}
\examples{
\dontrun {
  # Example of code in the app.R file
  library(surveydown)

  server <- function(input, output, session) {

    if (sd_is_answered("car_preference")) {
      # Do something only if the "car_preference" question is answered
    }

  }
}

}
