% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_next}
\alias{sd_next}
\title{Create a 'Next' Button for Page Navigation}
\usage{
sd_next(next_page = NULL, label = "Next")
}
\arguments{
\item{next_page}{Character string. The ID of the next page to navigate to. This parameter is required.}

\item{label}{Character string. The label of the 'Next' button. Defaults to "Next".}
}
\value{
A Shiny action button UI element.
}
\description{
This function creates a 'Next' button for navigating to the specified next page in a surveydown survey.
}
\details{
The function generates a Shiny action button that, when clicked, sets the input value
to the specified next page ID, facilitating page navigation within the Shiny application.
}
\examples{
\dontrun{
  ui <- fluidPage(
    sd_next(next_page = "page2", label = "Next Page")
  )
  server <- function(input, output, session) {
    observeEvent(input$next_page, {
      # Navigate to the next page
      shinyjs::hide("page1")
      shinyjs::show("page2")
    })
  }
  shinyApp(ui, server)
}

}
