% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_question}
\alias{sd_question}
\title{Create a survey question}
\usage{
sd_question(
  type,
  id,
  label,
  cols = "80",
  direction = "horizontal",
  status = "default",
  width = "100\%",
  height = "100px",
  selected = NULL,
  label_select = "Choose an option...",
  grid = TRUE,
  individual = TRUE,
  justified = FALSE,
  force_edges = TRUE,
  option = NULL,
  placeholder = NULL,
  required = FALSE,
  resize = NULL
)
}
\arguments{
\item{type}{Specifies the type of question. Possible values are "select", "mc", "mc_multiple", "mc_buttons", "mc_multiple_buttons", "text", "textarea", "numeric", "slider", "date", and "daterange". See "Details" section for more about each type.}

\item{id}{A unique identifier for the question, which will be used as the variable name in the resulting survey data.}

\item{label}{Character string. The label for the UI element, which can be formatted with markdown.}

\item{cols}{Integer. Number of columns for the textarea input. Defaults to 80.}

\item{direction}{Character string. The direction for button groups ("horizontal" or "vertical"). Defaults to "horizontal".}

\item{status}{Character string. The status for button groups. Defaults to "default".}

\item{width}{Character string. The width of the UI element. Defaults to "100\%".}

\item{height}{Character string. The height of the textarea input. Defaults to "100px".}

\item{selected}{Value. The selected value(s) for certain input elements.}

\item{label_select}{Character string. The label for the select input. Defaults to "Choose an option...".}

\item{grid}{Logical. Whether to show a grid for slider input. Defaults to TRUE.}

\item{individual}{Logical. Whether buttons in a group should be individually styled. Defaults to TRUE.}

\item{justified}{Logical. Whether buttons in a group should be justified. Defaults to FALSE.}

\item{force_edges}{Logical. Whether to force edges for slider input. Defaults to TRUE.}

\item{option}{List. Options for the select, radio, checkbox, and slider inputs.}

\item{placeholder}{Character string. Placeholder text for text and textarea inputs.}

\item{required}{Logical. Whether the input is required. Defaults to FALSE.}

\item{resize}{Character string. Resize option for textarea input. Defaults to NULL.}
}
\value{
A Shiny UI element wrapped in a div with a custom data attribute for question ID.
}
\description{
This function creates various types of survey questions.
}
\details{
Insert any detailed information here, such as more details on the types of questions
}
\examples{
# Insert examples here
}
