% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_question_reactive}
\alias{sd_question_reactive}
\title{Create a reactive survey question}
\usage{
sd_question_reactive(...)
}
\arguments{
\item{...}{Other inputs to be passed to sd_question.}
}
\value{
A renderUI function that creates the survey question.
}
\description{
This function creates various types of reactive survey questions.
}
