% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_redirect}
\alias{sd_redirect}
\title{Redirect button}
\usage{
sd_redirect(url, button = TRUE, label = "Click here", delay = NULL)
}
\arguments{
\item{url}{Character string. The target URL for redirection.}

\item{button}{Logical. If TRUE, creates a clickable button. If FALSE, creates non-clickable text. Defaults to TRUE.}

\item{label}{Character string. The text for the button or display. Defaults to "Click here".}

\item{delay}{Numeric. The delay in seconds before automatic redirection. If NULL, no automatic redirection occurs. Defaults to NULL.}
}
\value{
A Shiny UI element (button or text) with redirection functionality, horizontally centered and styled.
If a button is created, it includes JavaScript for Enter key functionality.
}
\description{
This function creates a UI element for redirecting to external links. It can provide a clickable button,
automatic redirection after a delay, or both. When a button is created, it can be activated by clicking
or by pressing the Enter key.
}
\details{
When \code{button = TRUE}, the function creates a clickable button that can be activated by mouse click or by pressing the Enter key.
The Enter key functionality is only active when the button is visible on the page.
If a delay is specified, the function will initiate an automatic redirection after the specified number of seconds.
}
\examples{
sd_redirect(
  url = "https://www.google.com",
  button = TRUE,
  label = "Go to Google",
  delay = 5)
sd_redirect(
  url = "https://www.example.com",
  button = FALSE,
  label = "Redirecting to Example",
  delay = 10
)

}
