% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_setup}
\alias{sd_setup}
\title{Required Set Up Function}
\usage{
sd_setup()
}
\value{
This function does not return a value. It is called for its side
effects of setting up the 'shiny' application.
}
\description{
This function is depreciated and no longer needed.
}
\details{
The function configures the 'shiny' application to use Bootstrap 5
for styling and enables 'shinyjs' for JavaScript functionalities within
the application.
}
