% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_show_password}
\alias{sd_show_password}
\title{Show the Saved Survey Password}
\usage{
sd_show_password()
}
\value{
A character string containing the password if found and confirmed,
or a message if no password is saved along with a suggestion to set one.
}
\description{
This function displays the password saved in the .Renviron file under the
SURVEYDOWN_PASSWORD variable. It includes a confirmation step to ensure
the user wants to display the password in the console. If no password is found,
it suggests using the sd_set_password() function to define a password.
}
\examples{
\dontrun{
  sd_show_password()
}

}
