% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_skip_if}
\alias{sd_skip_if}
\title{Define skip conditions for survey pages}
\usage{
sd_skip_if(...)
}
\arguments{
\item{...}{One or more formulas defining skip conditions.
The left-hand side of each formula should be a condition based on input
values, and the right-hand side should be the ID of the page to skip to if
the condition is met.}
}
\value{
A list of parsed conditions, where each element contains the
condition and the target page ID.
}
\description{
This function is used to define conditions under which certain pages in the
survey should be skipped. It takes one or more formulas where the left-hand
side is the condition and the right-hand side is the target page ID.
}
\examples{
\dontrun{
  # Example of implementing sd_skip_if in the Shiny server in the app.R file
  library(surveydown)

  server <- function(input, output, session) {

    # Assuming there are questions with the id values "age" and "country"
    # an pages with the id values "underage_page" and "international_page"

    sd_skip_if(
      input$age < 18 ~ "underage_page",
      input$country != "USA" ~ "international_page"
    )
  }
}

}
\seealso{
\code{sd_show_if()}
}
