% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_store_value}
\alias{sd_store_value}
\title{Store a custom value for the survey}
\usage{
sd_store_value(value, name = NULL)
}
\arguments{
\item{value}{The value to be stored.}

\item{name}{(Optional) The name for the value in the data.
If not provided, the name of the \code{value} variable will be used.}
}
\value{
NULL (invisibly)
}
\description{
This function allows storing additional values to be included in the survey data,
such as respondent IDs or other custom data. The values are stored in a special
environment (.sd_custom_values) and will be included when the survey data is saved.
}
\examples{
\dontrun{
  sd_store_value(respondentID)
  sd_store_value(respondentID, "respID")
}

}
