% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{sd_store_value}
\alias{sd_store_value}
\title{Store a value}
\usage{
sd_store_value(value, id = NULL)
}
\arguments{
\item{value}{The raid value to be stored.}

\item{id}{(Optional) The id (name) of the value in the data.
If not provided, the id of the \code{value} variable will be used.}
}
\value{
NULL (invisibly)
}
\description{
This function allows storing additional values to be included in the survey data,
such as respondent IDs or other data.
}
\examples{
\dontrun{
  sd_store_value(respondentID)
  sd_store_value(respondentID, "respID")
}

}
