% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_ui}
\alias{sd_ui}
\title{Create the UI for a surveydown survey}
\usage{
sd_ui(
  barcolor = "theme",
  barposition = "top",
  theme = "cosmo",
  backgroundcolor = "#f2f6f9",
  custom_css = NULL
)
}
\arguments{
\item{barcolor}{Color of the progress bar. Can be a hex color or 'theme' to use the theme's primary color.}

\item{barposition}{Position of the progress bar. Can be 'top', 'bottom', or 'none'.}

\item{theme}{The name of the Bootswatch theme to use.}

\item{backgroundcolor}{Background color of the body.}

\item{custom_css}{Path to a custom CSS file to override default styles.}
}
\value{
A Shiny UI object
}
\description{
This function creates the user interface for a surveydown survey,
including necessary CSS and JavaScript files, and applies custom styling.
}
