% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{sd_ui}
\alias{sd_ui}
\title{Create the UI for a surveydown survey}
\usage{
sd_ui(barcolor = NULL, barposition = "top")
}
\arguments{
\item{barcolor}{Color of the progress bar. Can be a hex color or NULL to use the theme's primary color with 80\% opacity. Default is NULL.}

\item{barposition}{Position of the progress bar. Can be 'top', 'bottom', or 'none'. Default is 'top'.}
}
\value{
A Shiny UI object
}
\description{
This function creates the user interface for a surveydown survey,
including necessary CSS and JavaScript files, and applies custom styling.
}
\examples{
\dontrun{
sd_ui()
sd_ui(barcolor = "#FF0000", barposition = "bottom")
}
}
