% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sd_update}
\alias{sd_update}
\title{Update Surveydown Package}
\usage{
sd_update()
}
\value{
No return value, called for side effects.
}
\description{
This function directly updates the surveydown package to the latest version
from GitHub, without checking the current version.
}
\examples{
\dontrun{
sd_update()
}
}
