% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{sd_update_extension}
\alias{sd_update_extension}
\title{Update Survey Extension}
\usage{
sd_update_extension(path = getwd())
}
\arguments{
\item{path}{A character string specifying the directory in which to update
or create the extension. Defaults to the current working directory.}
}
\value{
Invisibly returns TRUE if the extension was successfully updated.
}
\description{
This function updates or creates the _extensions/surveydown-dev/surveydown folder
with the latest contents from the surveydown-ext repository.
}
\details{
This function downloads the latest version of the surveydown extension from GitHub,
and updates the local copy in the specified path. If the extension directory
doesn't exist, it will be created.
}
\examples{
\dontrun{
sd_update_extension()
sd_update_extension(path = "path/to/survey")
}
}
